Polished FileSplashIcon
- Used one fill color instead of defining it per path

- Moved it to svgs directory
