+++ client/src/nj/svgs/FileSplashIcon.tsx
@@ -1,0 +1,27 @@
+import * as React from 'react';
+
+export default function FileSplashIcon() {
+  return (
+    <svg
+      width={88}
+      height={88}
+      viewBox="0 0 88 88"
+      fill="#1C1D1F"
+      xmlns="http://www.w3.org/2000/svg"
+      aria-hidden="true"
+    >
+      <path d="M19.7511 63.3842C16.8422 65.1393 14.2169 67.354 11.6502 69.5735C10.3693 70.6833 9.0933 71.798 7.8173 72.9126C6.62442 73.9589 5.36797 75.0295 4.75686 76.5451C3.57864 79.4638 6.51197 81.8349 9.24486 81.6002C10.8093 81.4633 12.1635 80.6175 13.3711 79.6691C14.7057 78.618 15.9769 77.4887 17.2333 76.3447C20.108 73.7242 22.8017 70.9278 25.3586 67.9944L24.1071 68.1558C29.8418 71.0989 37.2875 71.5389 42.5333 67.3295C47.212 63.5749 48.9818 56.6424 46.9333 51.0446C44.572 44.6011 37.9671 40.9002 31.4991 39.81C27.984 39.2184 24.2733 39.2722 20.8853 40.4798C17.776 41.5895 14.9453 43.6184 13.1217 46.4002C11.3177 49.1526 10.8191 52.5602 11.7969 55.7038C12.7746 58.8473 14.9893 61.5313 17.5462 63.5406C18.2111 64.0638 18.8955 64.5527 19.6044 65.0024C20.7826 65.7455 21.8582 63.8878 20.6897 63.1447C16.1382 60.2407 11.7626 54.7113 14.124 49.0598C16.5342 43.3006 23.3297 41.1642 29.0791 41.5944C34.9506 42.0295 41.2671 44.738 44.2444 50.0815C46.7231 54.5304 46.0093 60.5144 42.6262 64.3082C38.1773 69.3046 30.7755 69.07 25.1435 66.3713C24.7573 66.1855 24.2 66.166 23.892 66.5326C21.9217 68.8793 19.8195 71.1184 17.6244 73.25C16.5537 74.2913 15.4537 75.3082 14.3342 76.2958C13.2195 77.2784 12.1049 78.3442 10.8289 79.1166C9.84619 79.7082 8.70219 80.1189 7.58753 79.6495C6.56086 79.2193 6.40442 78.2758 6.83953 77.3126C7.40175 76.0806 8.56042 75.2153 9.54797 74.3402C10.648 73.3624 11.7235 72.3602 12.8137 71.3678C15.3217 69.0798 17.864 66.8309 20.5529 64.7531C21.3351 64.1518 20.6555 62.8366 19.7511 63.3793V63.3842Z" />
+      <path d="M18.568 48.5125C15.796 53.3378 19.58 58.5396 23.4862 61.336C26.7666 63.6827 31.1226 64.9147 35.1364 64.3867C37.1702 64.1178 39.116 63.3307 40.612 61.8982C42.2546 60.324 43.2422 58.1143 43.1395 55.8263C42.9293 51.0743 38.9937 47.2218 34.8968 45.3591C32.4133 44.2347 29.6657 43.6334 26.9328 43.7214C25.4857 43.7702 24.0435 44.0538 22.6991 44.5916C21.3253 45.144 19.976 45.8431 19.3942 47.2805C19.0128 48.2191 20.592 48.884 21.0368 47.9698C21.4817 47.0556 22.6844 46.6107 23.6133 46.2978C24.6888 45.936 25.8084 45.7454 26.9377 45.672C28.9862 45.54 31.0004 45.8725 32.9315 46.5667C36.5248 47.8622 40.2404 50.7271 41.0813 54.6431C41.8733 58.3391 39.3017 61.5071 35.7768 62.3138C32.4866 63.0667 28.7564 62.1036 25.8573 60.5C22.1515 58.4467 17.5462 53.7925 19.7657 49.2116C20.1471 48.4245 19.0226 47.74 18.5777 48.5174L18.568 48.5125ZM9.07373 81.2191C9.51862 82.3191 10.692 83.0574 11.8751 83.1063C13.1217 83.1551 14.6177 82.4951 15.2044 82.0503C16.0893 81.3805 16.8666 80.5836 17.6048 79.7525C20.6213 76.34 23.4177 72.7369 25.9746 68.9725C26.4391 69.432 27.016 69.7547 27.5977 70.0578C30.7804 71.7054 34.3493 72.8347 37.9231 72.5267C41.4968 72.2236 45.056 70.2827 46.6497 67.0707C47.5004 65.3547 47.6324 63.0227 46.2 61.7565C46.3662 62.744 45.8186 63.7022 45.2124 64.4991C41.6337 69.2071 35.1022 71.4463 29.3871 69.9209C28.0622 69.5689 26.7911 69.036 25.4564 68.7182C25.2022 68.6596 24.9284 68.6058 24.684 68.6938C24.508 68.7574 24.3662 68.8894 24.2293 69.0165C23.4422 69.7449 22.6551 70.4734 21.8631 71.2018C20.2351 72.7076 18.6022 74.2183 17.1502 75.9C16.5146 76.6383 15.9133 77.4107 15.224 78.1049C13.8876 79.457 12.2608 80.4869 10.4671 81.1165" />
+      <path d="M8.28183 81.4392C9.07871 83.385 11.0098 84.2601 13.0338 83.9032C15.092 83.5414 16.588 82.2312 17.9618 80.7352C19.5458 79.0045 21.0369 77.181 22.484 75.333C23.9312 73.485 25.2316 71.5196 26.6689 69.6619H25.2805C26.0334 70.4636 27.0845 70.9623 28.0623 71.4365C29.1378 71.9547 30.2427 72.4094 31.3818 72.7663C33.5867 73.4556 35.9774 73.8223 38.2849 73.5485C42.1325 73.0939 46.1805 70.7032 47.6765 67.0072C48.4636 65.0614 48.4587 62.6316 46.86 61.0916C46.3272 60.5783 45.1685 60.9059 45.2565 61.7565C45.3934 63.0667 44.264 64.2254 43.4085 65.0956C42.5236 65.9952 41.5263 66.7579 40.436 67.3885C38.3143 68.6156 35.8992 69.3196 33.4498 69.3881C32.2227 69.4223 31.0103 69.2952 29.8223 68.9921C29.0645 68.7965 28.3214 68.5374 27.5832 68.2881C26.7863 68.0192 25.9112 67.6721 25.0654 67.6232C24.2196 67.5743 23.7112 68.073 23.1489 68.5912C22.528 69.1632 21.9072 69.7352 21.2912 70.3072C20.108 71.3974 18.9249 72.4827 17.8005 73.6365C16.544 74.9272 15.5027 76.4134 14.1925 77.6503C13.0394 78.7343 11.6768 79.5712 10.1885 80.1094C9.6556 80.3001 9.29871 80.8232 9.45516 81.3952C9.59694 81.9085 10.208 82.3192 10.7409 82.1285C12.2614 81.581 13.6547 80.8183 14.9112 79.8014C16.1676 78.7845 17.1258 77.4987 18.1916 76.3059C19.3112 75.0494 20.5236 73.881 21.7556 72.7321C22.3716 72.1552 22.9925 71.5881 23.6085 71.0112C23.892 70.7521 24.1707 70.4881 24.4543 70.229C24.6205 70.0725 24.7867 69.9063 24.9627 69.7596C25.1289 69.6179 25.0556 69.6472 25.2952 69.7352C26.0285 70.0187 26.8352 70.185 27.588 70.4343C28.38 70.6934 29.172 70.9427 29.9885 71.0992C31.3036 71.3583 32.6529 71.4561 33.9876 71.373C36.7205 71.2019 39.4192 70.3707 41.7463 68.9236C43.0027 68.1414 44.1663 67.1881 45.1538 66.0881C46.3174 64.7876 47.3489 63.3405 47.0996 61.5121L45.5303 62.4263C46.6303 63.4236 46.508 65.1201 45.9556 66.3814C45.3249 67.8285 44.176 69.041 42.8609 69.8916C39.8445 71.8374 35.9334 71.8521 32.5747 70.9281C31.5285 70.6396 30.5067 70.2681 29.5094 69.833C28.5365 69.4076 27.4903 68.9676 26.6689 68.2832C26.1898 67.8823 25.476 67.897 25.1289 68.4836C24.0925 70.2339 22.8067 71.8716 21.5698 73.4899C20.328 75.113 19.0325 76.7019 17.6978 78.2467C16.588 79.5374 15.4049 81.1019 13.7965 81.7863C12.3445 82.4023 10.56 82.4267 9.78271 80.8085C9.38671 79.9823 7.92494 80.5739 8.28183 81.4392Z" />
+      <path d="M7.73917 75.6411C9.35739 76.6628 10.7996 77.9437 12.0463 79.4593C12.2907 79.7575 12.8383 79.7331 13.0876 79.4593C13.3761 79.1464 13.3516 78.7308 13.0876 78.4179C11.7676 76.8486 10.1836 75.5237 8.41384 74.4873C8.10095 74.3064 7.6805 74.414 7.49962 74.7268C7.31873 75.0397 7.43117 75.4504 7.73917 75.6411ZM10.8876 72.3264C11.1076 72.346 11.4009 72.5904 11.5721 72.7273C11.8703 72.9717 12.1538 73.2308 12.4325 73.4997C12.9852 74.0233 13.5091 74.5766 14.0018 75.1571C14.2365 75.4357 14.4272 75.6802 14.6276 75.954C14.8672 76.2815 15.0676 76.6824 15.3463 76.9757C15.7961 77.4451 16.6467 77.3668 16.8325 76.6726C16.9889 76.086 16.2996 75.3966 15.7325 75.8073C15.4294 76.0273 15.2485 76.5357 15.6005 76.8291L15.6201 76.8437L15.4245 76.3695V76.3939C15.4587 76.2766 15.4881 76.1593 15.5223 76.0468C15.6054 75.9442 15.6201 75.9246 15.5663 75.9833L15.8987 75.7926L15.8596 75.8024H16.2605C16.4952 76.0224 16.5685 76.0762 16.4854 75.954C16.4609 75.9148 16.4365 75.8806 16.4121 75.8415L16.2849 75.6557C16.2018 75.5335 16.1138 75.4113 16.0258 75.2891C15.8205 75.0055 15.6103 74.7317 15.3903 74.458C14.9992 73.9691 14.5787 73.4997 14.1338 73.0597C13.6889 72.6197 13.2294 72.2286 12.7503 71.8522C12.2027 71.422 11.6209 71.0211 10.8925 71.0651C10.5552 71.0846 10.2667 71.3388 10.2667 71.6908C10.2667 72.0428 10.5552 72.2824 10.8925 72.3166L10.8876 72.3264ZM14.0898 69.1731C14.4565 69.5886 14.8867 69.9553 15.2925 70.3317C15.6983 70.7082 16.1138 71.1286 16.5049 71.5442C16.8961 71.9597 17.2578 72.3704 17.6147 72.8006C17.7761 72.9913 17.9814 73.2553 18.1036 73.4068C18.3334 73.7051 18.5827 74.1108 19.0765 74.1108C19.3943 74.1108 19.6778 73.8908 19.7658 73.5877C19.9076 73.094 19.5801 72.7664 19.3014 72.4095C18.764 71.7311 18.1929 71.0801 17.5903 70.4588C17.1894 70.0482 16.7836 69.6473 16.3583 69.2611C15.9085 68.8504 15.4245 68.4251 14.8965 68.1219C14.6129 67.9557 14.2365 68.0291 14.0409 68.2931C13.8454 68.5571 13.8698 68.9091 14.0947 69.1633L14.0898 69.1731ZM22.6649 49.862C23.0267 51.2308 23.6183 52.5704 24.1854 53.866C24.7525 55.1615 25.3734 56.4375 26.1165 57.6353C26.5858 58.3931 27.6956 57.7282 27.3045 56.9411C26.6934 55.7188 26.0676 54.5064 25.5054 53.2548C24.9432 52.0033 24.513 50.742 23.9947 49.4953C23.853 49.1531 23.5401 48.9086 23.1489 49.0162C22.7872 49.114 22.5721 49.5002 22.6698 49.862H22.6649ZM28.6734 48.542C30.013 51.9837 31.3965 55.4793 33.1712 58.7255C33.6161 59.542 34.7405 58.8233 34.3934 58.0117C32.9756 54.7166 31.4258 51.49 30.1058 48.146C29.9592 47.7695 29.6121 47.5202 29.1965 47.6277C28.8396 47.7206 28.5365 48.1655 28.6783 48.5371L28.6734 48.542ZM37.5076 57.0486C38.4805 57.0486 38.4805 55.538 37.5076 55.538C36.5347 55.538 36.5347 57.0486 37.5076 57.0486ZM34.5254 40.5437C33.2152 37.3855 31.9441 34.2664 31.0787 30.9517L30.1401 32.1886C32.8192 32.1446 35.4934 32.1202 38.1725 32.1788C38.813 32.1935 39.4534 32.2082 40.0889 32.2326C40.4947 32.2473 41.3258 32.1251 41.6681 32.4184C41.937 32.6482 42.0445 33.1566 42.1765 33.4695C42.3769 33.9437 42.6263 34.3837 42.993 34.7504C44.4938 36.2415 46.8649 36.1046 48.8107 36.1193C51.5192 36.1388 54.2227 36.1926 56.9312 36.2904C58.4663 36.3442 60.0014 36.4077 61.5316 36.486C62.2601 36.5202 62.9836 36.5593 63.7121 36.6033L64.9245 36.6766C65.1787 36.6913 65.5747 36.6375 65.7801 36.7842L65.4281 36.4322C65.7263 36.8331 65.8094 37.5028 65.9609 37.9771L66.5476 39.8397C66.9143 41.0033 67.281 42.1717 67.6525 43.3353C68.4152 45.7504 69.2463 48.1508 69.9161 50.6002C71.0698 54.7948 71.4072 59.5322 73.9592 63.1891L74.8098 62.0842C72.0721 61.6246 69.3441 61.0868 66.6063 60.6224C62.485 59.9184 58.3392 59.3268 54.2032 58.7255C51.8125 58.3784 49.4218 58.0117 47.0263 57.6793C46.4934 57.606 45.9556 57.8064 45.7992 58.3735C45.6672 58.8575 45.9654 59.5028 46.4934 59.6006C50.6881 60.3731 54.9316 60.8913 59.1556 61.4535C62.8858 61.9473 66.6161 62.4362 70.3463 62.9202C71.7054 63.0962 73.0694 63.2477 74.4089 63.5362C75.0689 63.678 75.553 63.0277 75.2596 62.4313C73.3872 58.6424 72.8396 54.5015 71.8178 50.4486C70.6494 45.8286 69.0752 41.2966 67.6476 36.75C67.5009 36.2757 67.3983 35.7184 67.0609 35.3322C66.6845 34.902 66.2054 34.7944 65.6578 34.7455C64.4503 34.638 63.2378 34.5988 62.0254 34.5353C59.6836 34.4131 57.3418 34.3202 55.0001 34.2566C52.5752 34.1882 50.1503 34.1588 47.7303 34.1442C46.3027 34.1344 44.5721 34.1148 43.9463 32.5944C43.5356 31.5922 43.0712 30.6731 41.9223 30.4042C40.7392 30.1304 39.3898 30.2428 38.1823 30.2135C35.5032 30.1548 32.8241 30.15 30.1498 30.2526C29.5241 30.2771 29.0303 30.8588 29.2112 31.4895C30.1547 34.8091 31.4258 38.1726 33.2885 41.0913C33.6845 41.7073 34.8481 41.3113 34.5401 40.5633L34.5254 40.5437Z" />
+      <path d="M38.2898 30.2229C37.6347 26.6002 36.6325 23.0313 35.3614 19.5798C35.0583 19.9758 34.76 20.3718 34.4569 20.7629C44.7578 20.802 55.0636 20.714 65.3645 20.6847C65.9658 20.6847 66.5671 20.6749 67.1685 20.6847C67.6085 20.6847 68.1365 20.6602 68.4885 20.9682C69.1729 21.5647 69.3294 22.8896 69.5347 23.7109C70.1116 26.0478 70.4147 28.4384 70.6396 30.834C71.1089 35.8304 71.4218 40.8367 71.808 45.838C72.0231 48.6589 72.2676 51.4798 72.5756 54.2958C72.732 55.7184 72.908 57.1362 73.1134 58.554C73.2063 59.214 73.2747 59.8984 73.4507 60.5438C73.656 61.2967 73.7343 62.0251 73.8174 62.8024C73.9396 63.9318 75.6067 63.7069 75.3574 62.5922C75.2156 61.9664 75.0298 61.326 75.0102 60.6856C74.9858 60.0109 74.9271 59.3509 74.8489 58.6811C74.6974 57.3807 74.5654 56.0753 74.4431 54.77C74.1987 52.1936 74.0129 49.6122 73.8418 47.0309C73.4947 41.8584 73.172 36.686 72.6978 31.5233C72.4631 28.9909 72.1845 26.4536 71.6271 23.97C71.1774 21.946 70.6787 19.1691 68.2245 18.7976C66.9778 18.6069 65.6529 18.734 64.3965 18.7389L60.3827 18.7584C57.7476 18.7731 55.1076 18.7829 52.4725 18.7976C46.4689 18.8269 40.4654 18.8269 34.4667 18.8953C33.9094 18.9002 33.3374 19.4967 33.5623 20.0784C34.892 23.5251 35.9383 27.05 36.7791 30.6433C37.0138 31.6407 38.4756 31.2104 38.2947 30.2229H38.2898Z" />
+      <path d="M49.808 25.8526C49.6173 25.4126 49.4902 24.8162 49.7151 24.3664C49.8911 24.0144 50.3262 23.7358 50.7222 23.8922C51.5337 24.21 51.7391 25.6278 51.8711 26.3464C51.964 26.85 52.052 27.3633 52.052 27.8766C52.052 28.39 51.9053 28.6638 51.392 28.7958C50.5266 29.0206 49.4804 29.138 48.5906 29.006C47.5542 28.8544 47.0115 28.2238 46.6351 27.29C46.1853 26.1802 45.8724 24.9971 45.6475 23.8238C45.4226 22.6504 45.3102 21.4184 45.3102 20.2011C45.3102 19.2086 45.0902 17.6686 46.0728 17.0869C47.0555 16.5051 48.3657 16.1775 49.5 16.3193C50.6048 16.4562 50.8395 18.2015 50.9226 19.0962C50.9764 19.6633 51.3675 20.1522 51.9786 20.1522C52.5066 20.1522 53.0884 19.6682 53.0346 19.0962C52.888 17.5415 52.4431 15.3318 50.8884 14.5544C49.4364 13.8309 47.4124 14.3344 46.0142 14.9504C45.3004 15.2633 44.5573 15.674 44.1124 16.3389C43.6137 17.082 43.5942 17.9718 43.5893 18.8371C43.5746 20.6509 43.7653 22.4695 44.1026 24.2491C44.2591 25.0704 44.4497 25.8771 44.6844 26.6789C44.9191 27.4806 45.1342 28.3362 45.5888 29.05C46.4395 30.3895 47.9013 30.8344 49.4217 30.7855C50.6831 30.7415 52.5262 30.6584 53.3231 29.5144C54.1982 28.258 53.6408 26.1362 53.1813 24.8162C52.7217 23.4962 51.5924 21.9562 49.9448 22.3913C48.136 22.8753 47.6373 25.0998 48.6786 26.5126C49.1822 27.202 50.1453 26.6251 49.8128 25.8478L49.808 25.8526ZM56.2124 19.3798C56.5644 15.6006 56.8431 11.7775 57.4737 8.03753L56.3591 8.6682C60.3337 9.2402 64.3866 9.45042 68.3906 9.70953C72.4044 9.96864 76.4231 10.1593 80.4466 10.2766C80.1582 9.77309 79.8697 9.27442 79.5813 8.77086C76.8142 13.8015 75.3866 19.2331 74.4284 24.8602C73.4213 30.766 72.5022 36.7744 72.468 42.778C72.4484 46.21 72.6782 49.6273 72.5071 53.0593C72.4533 54.1495 74.14 54.1446 74.2035 53.0593C74.3697 50.1162 74.3697 47.1584 74.404 44.2104C74.4382 41.2478 74.6142 38.3095 74.9417 35.3664C75.5724 29.6709 76.3791 23.8775 77.924 18.3433C78.7502 15.3855 79.8306 12.4718 81.3071 9.77309C81.6737 9.09842 81.2093 8.29175 80.4417 8.26731C76.4622 8.15486 72.4826 7.98375 68.508 7.7442C64.5382 7.50464 60.5684 7.06953 56.5986 6.87886C56.0853 6.85442 55.7284 7.31397 55.6942 7.78331C55.4155 11.6504 54.9217 15.498 54.6528 19.3651C54.5844 20.3722 56.1195 20.3526 56.2124 19.3651V19.3798Z" />
+      <path d="M75.1716 31.7583C77.6013 31.8756 80.0214 32.0614 82.4413 32.3156C82.1969 32.6139 82.0551 32.9707 81.9133 33.3276C78.5742 41.9174 76.2227 52.2525 74.7707 61.3507C74.6142 62.3236 75.0249 61.5023 74.5262 62.353C74.2378 61.1992 74.7316 60.7396 74.6289 59.5565L73.9249 51.5779C73.7734 49.8716 73.6218 48.1459 73.8907 46.4592C73.9982 45.7845 74.1693 45.1245 74.3307 44.4645C75.3133 40.4019 75.8853 36.2365 76.0271 32.0565" />
+      <path d="M75.1716 32.5306C77.5964 32.731 80.0262 32.9999 82.4413 33.2835C82.1627 32.7995 81.884 32.3106 81.6053 31.8266C80.9796 32.643 80.6764 33.655 80.3196 34.6084C79.9627 35.5617 79.6204 36.5346 79.2978 37.5075C78.6311 39.5422 78.0197 41.5947 77.4644 43.6626C76.3302 47.867 75.4062 52.1253 74.6044 56.4079C74.3893 57.5519 74.1889 58.7008 73.9933 59.8497C73.9004 60.407 73.7782 60.9693 73.7293 61.5315C73.7147 61.727 73.7147 61.9177 73.7684 62.1133C73.8027 62.2306 74.1791 62.9102 73.9053 62.4164L73.8027 61.6342C73.8222 61.5364 73.964 61.3702 73.832 61.5266C73.7587 61.6097 73.7 61.7466 73.6462 61.8395C74.2622 61.9177 74.8782 62.0008 75.4893 62.079C75.328 61.3653 75.6164 60.6808 75.6262 59.967C75.6409 58.9942 75.4649 57.987 75.3769 57.0142C75.1911 54.9315 74.9907 52.8439 74.8196 50.7613C74.7413 49.7933 74.6778 48.8204 74.7169 47.8475C74.756 46.8746 74.9662 45.9066 75.2009 44.9435C75.68 42.9928 76.076 41.0226 76.3644 39.0377C76.7067 36.7253 76.9267 34.3982 77.0196 32.0613C77.0684 30.7608 75.0396 30.7608 74.9956 32.0613C74.8636 36.1582 74.3209 40.2208 73.3676 44.2053C72.8884 46.2097 72.6733 48.1213 72.8151 50.1844C72.9569 52.2475 73.1769 54.423 73.3627 56.5399C73.4604 57.6693 73.6609 58.8426 73.6316 59.9719C73.6218 60.3239 73.5289 60.6613 73.4751 61.0084C73.392 61.5657 73.4213 62.0742 73.5387 62.6217C73.7244 63.4968 74.9564 63.5702 75.3818 62.8613C75.5138 62.6462 75.68 62.4262 75.7436 62.1768C75.812 61.9177 75.7729 61.6195 75.636 61.3848C75.548 61.2382 75.3573 60.9057 75.6067 61.4728C75.8951 62.1279 75.6751 61.727 75.7338 61.6195C75.8364 61.4435 75.8169 61.111 75.8511 60.9057L75.9684 60.187C76.0613 59.6346 76.1542 59.0773 76.252 58.5248C76.6284 56.3933 77.044 54.2666 77.5084 52.1546C78.4471 47.8573 79.5422 43.5942 80.8818 39.4044C81.2484 38.2506 81.6396 37.1066 82.0404 35.9626C82.2262 35.4346 82.412 34.9115 82.6027 34.3835C82.7933 33.8555 82.9351 33.259 83.2627 32.8093C83.732 32.1542 83.0818 31.411 82.4267 31.3524C80.0018 31.1324 77.5916 31.0346 75.1569 30.9955C74.1498 30.9808 74.1791 32.4573 75.1569 32.5404L75.1716 32.5306ZM61.5707 13.4248C63.0618 13.6888 64.5578 13.8844 66.0684 14.0115C66.8067 14.0702 67.5449 14.1093 68.288 14.1337C69.0116 14.1533 69.9209 14.2999 70.6151 14.0848C71.3093 13.8697 71.3044 12.9017 70.6151 12.7013C70.224 12.5888 69.8036 12.6328 69.3978 12.6279C69.0556 12.6279 68.7133 12.623 68.3711 12.6133C67.6329 12.5986 66.8898 12.5693 66.1516 12.535C64.6751 12.4666 63.1987 12.359 61.732 12.2124C61.0427 12.1439 60.8813 13.3075 61.5658 13.4297L61.5707 13.4248ZM62.0987 16.7444C62.8711 16.891 63.6582 16.9742 64.4404 17.0524L65.5209 17.155C65.9364 17.1942 66.5231 17.3555 66.88 17.0866C67.2564 16.8079 67.364 16.3337 67.0267 15.967C66.6893 15.6004 66.0733 15.6542 65.6089 15.615L64.5284 15.527C63.7804 15.4635 63.0324 15.4097 62.2844 15.4146C61.9471 15.4146 61.6636 15.6688 61.6196 16.0013C61.5804 16.3093 61.7711 16.6808 62.1036 16.7444H62.0987Z" />
+      <path d="M45.8382 62.1331C49.6858 67.3495 53.8756 72.3313 57.948 77.3717C58.2365 77.7286 58.8085 77.7237 59.18 77.5282C65.7702 74.0424 73.4507 72.9864 79.6693 68.7282C80.2462 68.3322 80.3196 67.6526 79.8747 67.1246C78.3151 65.2668 76.7556 63.4091 75.1813 61.566C74.8293 61.1553 74.1253 61.2042 73.7685 61.566C73.3578 61.9766 73.4213 62.5633 73.7685 62.9788C75.3133 64.8464 76.8729 66.7091 78.4324 68.5668L78.6378 66.9633C72.4289 71.2166 64.724 72.2384 58.1533 75.7682C58.564 75.8219 58.9747 75.8757 59.3853 75.9246C55.3325 70.9771 51.3431 65.9317 47.0458 61.1944C46.4689 60.5588 45.3151 61.4144 45.8382 62.1233V62.1331Z" />
+      <path d="M54.4035 62.7197C55.4448 62.4557 56.4862 62.1917 57.5324 61.9277C58.0555 61.7957 58.5737 61.6637 59.0968 61.5268C59.6199 61.3899 60.1626 61.3019 60.6515 61.1015C61.4728 60.769 61.1648 59.6397 60.2848 59.7766C59.7617 59.8548 59.2386 60.0601 58.7302 60.2068C58.2119 60.3535 57.6986 60.5099 57.1853 60.6615C56.1537 60.9646 55.1222 61.2726 54.0955 61.5757C53.3671 61.7908 53.6702 62.9055 54.4084 62.7197H54.4035ZM57.7084 65.8046C58.8768 66.2006 60.1284 65.4721 60.5244 64.3428L59.3657 64.6459C59.7422 65.1104 60.2702 65.5161 60.9008 65.477C61.5022 65.4428 61.9666 65.0175 62.5093 64.8121C63.1008 64.5921 63.7119 64.4113 64.3182 64.2157C64.6164 64.1179 64.9097 64.025 65.208 63.9273C65.5551 63.8148 65.9462 63.7317 66.2444 63.5117C66.7577 63.1304 66.6257 62.2504 65.9022 62.2455C65.3351 62.2455 64.7582 62.5241 64.2253 62.705C63.7071 62.8761 63.1937 63.0473 62.6755 63.2233C62.1573 63.3993 61.6831 63.5704 61.2284 63.8295C61.1355 63.8833 60.9155 64.0544 60.8471 64.0544C60.6564 64.0544 60.4511 63.7855 60.3435 63.6633C59.9915 63.2721 59.3168 63.4824 59.1848 63.9664C59.0528 64.4504 58.5297 64.7095 58.0457 64.5579C57.2537 64.3086 56.9262 65.5357 57.7035 65.7948L57.7084 65.8046ZM61.5266 69.1046C62.9688 68.6206 64.4159 68.1561 65.8679 67.7064C66.5817 67.4864 67.2955 67.2664 68.0093 67.0513C68.3808 66.9388 68.7524 66.8264 69.1239 66.7188C69.4955 66.6113 69.9208 66.5428 70.2386 66.3081C70.8155 65.8779 70.6346 64.9881 69.8671 64.9393C69.4662 64.9148 69.0848 65.0859 68.7035 65.2033C68.3222 65.3206 67.9653 65.433 67.5937 65.5504C66.8799 65.7753 66.1711 66.0099 65.4671 66.2495C64.0248 66.7384 62.5924 67.2468 61.1599 67.7699C60.3337 68.073 60.6759 69.3881 61.5266 69.1046Z" />
+      <path d="M47.3684 63.9712C50.0817 70.0236 52.8 76.0859 55.9875 81.9085C56.0169 81.9623 56.0511 82.0161 56.1049 82.0356C56.1586 82.0552 56.2173 82.0356 56.2662 82.0161C57.6057 81.5076 58.9453 80.9992 60.28 80.4907C65.3937 78.545 70.5124 76.5992 75.4013 74.1401C74.249 73.6619 73.2036 72.9589 72.3262 72.0721C70.7764 72.7174 69.2217 73.3579 67.672 74.0032C65.2373 75.0103 62.8075 76.0174 60.324 76.9023C59.8644 77.0685 59.3657 77.2299 58.8964 77.0979C58.4613 76.9805 58.1289 76.6383 57.8257 76.301C55 73.1574 52.7804 69.3441 49.2115 67.0756" />
+      <path d="M46.6155 64.1766C48.6688 69.3637 51.1524 74.3601 53.592 79.3761C53.8804 79.9726 54.1786 80.569 54.4817 81.1606C54.7604 81.7032 55 82.5001 55.5328 82.8423C56.1391 83.2383 56.8235 82.8717 57.4297 82.6517C58.0995 82.4072 58.7742 82.1579 59.444 81.9037C62.0448 80.9161 64.6506 79.9383 67.2271 78.9019C70.1751 77.7188 73.1133 76.4966 75.9537 75.0788C76.6382 74.7366 76.7017 73.5046 75.9537 73.1868C74.8831 72.7272 73.9346 72.1063 73.1035 71.285C72.8493 71.0357 72.3848 70.8548 72.0328 71.0014C68.3075 72.5414 64.6017 74.1548 60.8177 75.5481C60.3973 75.7046 59.9671 75.9001 59.5271 75.993C59.1408 76.0761 58.9991 75.9392 58.7302 75.6606C58.1435 75.0494 57.6106 74.3797 57.0826 73.7246C56.056 72.4534 55.0733 71.1432 53.9977 69.9112C52.7462 68.4739 51.3626 67.1637 49.7591 66.1321C48.5808 65.3743 47.4955 67.2517 48.664 68.0046C51.5533 69.8721 53.5333 72.7908 55.66 75.421C56.8186 76.8583 58.0066 78.6379 60.1088 78.105C61.0817 77.8606 62.0351 77.4303 62.9688 77.0734C63.976 76.6872 64.9733 76.2863 65.9706 75.8757C68.1902 74.9663 70.4048 74.0423 72.6195 73.1232L71.5488 72.8397C72.5022 73.7832 73.612 74.5508 74.8488 75.0788V73.1868C71.7884 74.7268 68.6595 76.1103 65.4768 77.3717C63.8782 78.0072 62.2746 78.6232 60.6662 79.2343C59.8888 79.5326 59.1115 79.8259 58.3342 80.1241C57.5568 80.4223 56.6671 80.9406 55.836 81.0677L56.6035 81.1703L56.5351 81.1214L56.892 81.4783C56.0022 80.1388 55.2688 78.6281 54.5453 77.1957C53.7875 75.6997 53.0542 74.189 52.3306 72.6734C50.8933 69.6472 49.544 66.5721 48.0382 63.5752C47.6471 62.7979 46.288 63.3454 46.6155 64.1766ZM31.108 13.4152C26.7862 12.8921 22.3128 14.5788 19.5115 17.9374C18.1426 19.5801 17.1404 21.5161 16.676 23.6086C16.4413 24.6646 16.2262 25.8623 16.4608 26.9379C16.6955 28.0134 17.468 28.7566 18.3382 29.3139C21.3253 31.2108 25.7155 30.365 27.9351 27.6614C28.5657 26.8939 29.3235 25.8086 29.0448 24.7672C28.7662 23.7259 27.5488 23.3446 26.576 23.2126C24.6008 22.9437 22.5768 23.545 20.9488 24.6646C16.6173 27.6468 15.6395 33.1566 15.9328 38.0797C16.0062 39.2726 17.776 39.2921 17.8053 38.0797C17.8835 34.6134 18.0497 30.5899 20.4404 27.8277C21.4573 26.6494 22.7773 25.7206 24.2977 25.3343C24.9626 25.1632 25.6666 25.1094 26.3511 25.2072C26.5466 25.2366 26.7568 25.2757 26.9426 25.3392C26.9671 25.349 27.0355 25.3734 27.0697 25.3783C27.3044 25.4077 27.0648 25.5152 27.148 25.349C27.3093 25.0312 26.9915 25.3881 27.0502 25.349C27.0111 25.3783 26.9426 25.5641 26.9084 25.6228C26.8204 25.7792 26.7275 25.9308 26.6248 26.0774C25.1973 28.121 22.0635 28.9521 19.8244 27.8423C19.228 27.549 18.5191 27.1237 18.3822 26.4148C18.2306 25.6374 18.4408 24.7134 18.6168 23.9557C18.9933 22.3326 19.7022 20.7974 20.7044 19.4677C23.1048 16.2703 27.1626 14.6081 31.1128 14.9552C32.1053 15.0432 32.0662 13.5423 31.1128 13.4299L31.108 13.4152ZM38.9106 9.88056C38.632 9.70945 38.368 9.52856 38.1088 9.33301L38.2653 9.45523C38.1431 9.35745 38.0208 9.25967 37.8986 9.15701L37.7226 9.01034L37.5662 8.88812L37.3804 8.80989L37.3364 8.79523C37.2973 8.78056 37.2582 8.77078 37.2191 8.75612L37.0333 8.73167L36.8475 8.75612C36.7008 8.79523 36.5737 8.88812 36.4808 9.01034C36.4026 9.11301 36.3586 9.21567 36.344 9.34278C36.3293 9.40634 36.3293 9.46501 36.344 9.52856C36.344 9.62145 36.3782 9.70945 36.4271 9.78767C36.4857 9.91967 36.5346 10.0223 36.6422 10.1299C36.696 10.1837 36.7546 10.2326 36.8133 10.2815C36.9844 10.4183 37.1604 10.5503 37.3364 10.6775C37.5955 10.8632 37.8644 11.0343 38.1431 11.1908C38.3142 11.2886 38.5342 11.3179 38.7248 11.269C38.9008 11.2201 39.0915 11.0832 39.1746 10.9219C39.2675 10.7459 39.3115 10.5357 39.2528 10.3401C39.1942 10.1446 39.0768 9.99301 38.9057 9.89034L38.9106 9.88056ZM45.3004 8.03745C45.0069 6.9455 44.6461 5.87278 44.22 4.82545C44.1368 4.62501 44.044 4.47345 43.8582 4.35612C43.648 4.21923 43.3791 4.22412 43.1591 4.31701C42.6995 4.51256 42.5822 5.00634 42.7826 5.43167C42.9048 5.68589 43.0173 5.94501 43.1297 6.20412C43.1982 6.36545 43.0857 6.09656 43.1493 6.24812C43.1786 6.32145 43.2128 6.39967 43.2422 6.47301C43.2911 6.59523 43.34 6.71745 43.384 6.83967C43.5746 7.36278 43.7457 7.89078 43.8973 8.42367C44 8.79034 44.4155 9.05434 44.792 8.93212C45.1684 8.80989 45.4031 8.43345 45.3004 8.03745ZM51.6951 6.73212V6.69789C51.6951 6.76145 51.7097 6.82989 51.7146 6.89345C51.6755 6.60012 51.6804 6.31167 51.7146 6.01834L51.6853 6.21878C51.7391 5.83745 51.8466 5.45612 51.9884 5.09434C52.052 4.92812 52.008 4.65434 51.9102 4.50767C51.8124 4.36101 51.7 4.26323 51.5435 4.19478C51.2208 4.05789 50.7515 4.16056 50.5902 4.50767C50.4924 4.71789 50.3897 4.93789 50.3311 5.16767C50.2724 5.39745 50.2431 5.63701 50.2186 5.87167C50.1746 6.28723 50.2137 6.70767 50.2773 7.11834C50.3017 7.28945 50.4728 7.47523 50.6146 7.55834C50.7711 7.65123 51.0057 7.69034 51.1817 7.63167C51.3577 7.57301 51.5288 7.46545 51.6217 7.29434L51.6951 7.11834C51.7293 6.98634 51.7293 6.85923 51.6951 6.72723V6.73212Z" />
+    </svg>
+  );
+}
